+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/Veining.java
@@ -193,7 +193,7 @@ public class Veining extends Enchantment {
     }
 
     public static int getAmountMined(int lvl) {
-        return (lvl + 1) * (lvl + 1);
+        return lvl * 4;
     }
 
     public static List<BlockPos> getMinedBlocks(ItemStack heldStack, int lvl, Level level, LivingEntity entity, BlockPos targetPos, Direction face) {